+++ src/CleanRegex/Builder/PatternBuilder.php
@@ -8,7 +8,7 @@ use TRegx\CleanRegex\Internal\Prepared\Parser\MaskParser;
 use TRegx\CleanRegex\Internal\Prepared\Parser\PreparedParser;
 use TRegx\CleanRegex\Internal\Prepared\PrepareFacade;
 use TRegx\CleanRegex\Internal\Prepared\Template\NoTemplate;
-use TRegx\CleanRegex\PatternInterface;
+use TRegx\CleanRegex\Pattern;
 use TRegx\CleanRegex\Template;
 
 class PatternBuilder
@@ -18,22 +18,22 @@ class PatternBuilder
         return new PcrePatternBuilder();
     }
 
-    public function bind(string $input, array $values, string $flags = null): PatternInterface
+    public function bind(string $input, array $values, string $flags = null): Pattern
     {
         return PrepareFacade::build(new BindingParser($input, $values, new NoTemplate()), new StandardStrategy($flags ?? ''));
     }
 
-    public function inject(string $input, array $values, string $flags = null): PatternInterface
+    public function inject(string $input, array $values, string $flags = null): Pattern
     {
         return PrepareFacade::build(new InjectParser($input, $values, new NoTemplate()), new StandardStrategy($flags ?? ''));
     }
 
-    public function prepare(array $input, string $flags = null): PatternInterface
+    public function prepare(array $input, string $flags = null): Pattern
     {
         return PrepareFacade::build(new PreparedParser($input), new StandardStrategy($flags ?? ''));
     }
 
-    public function mask(string $mask, array $keywords, string $flags = null): PatternInterface
+    public function mask(string $mask, array $keywords, string $flags = null): Pattern
     {
         return PrepareFacade::build(new MaskParser($mask, $keywords), new StandardStrategy($flags ?? ''));
     }