+++ test/Utils/AssertsPattern.php
@@ -2,11 +2,11 @@
 namespace Test\Utils;
 
 use PHPUnit\Framework\Assert;
-use TRegx\CleanRegex\PatternInterface;
+use TRegx\CleanRegex\Pattern;
 
 trait AssertsPattern
 {
-    public function assertSamePattern(string $expected, PatternInterface $actual): void
+    public function assertSamePattern(string $expected, Pattern $actual): void
     {
         Assert::assertSame($expected, $actual->delimited());
     }